refactor(container): move BindToFluentSyntax
